修正
`set_commentable`を`CommentsController`から削除して条件分岐をせずに済むように`Books::CommentsController`と`Reports::CommentsController`に移動。
`create`メソッドの`elseif`を削除し、`else`に変更。`else`内に`render_commentable_show`メソッドを使用。
`render_commentable_show`は`Books::CommentsController`、`Reports::CommentsController`に定義。
`destroy`メソッドを削除失敗時に例外が発生するように`destroy!`に変更。
`user_confirmation`を動詞から始まるようにメソッド名を変更。
使用していないビューヘルパーを削除。
`users_helper.rb`に作成した`name_or_email`メソッドを`user.rb`に移動。
移動した`name_or_email`が全角スペースだけのユーザー名には対応できてなかったので修正。
各ビューファイルにエラーメッセージのコードが散らばっていたのを`views/shared/form_error_messages.html.erb`に統一して`render`で呼び出せるように修正。
books,reportsのshowページが表示されるように部分テンプレート名を変更。
モデルファイルに移動させた`name_or_email`メソッドをビューで使えるように修正。
`books`と`reports`のルーティングを名前空間付きコントローラー`Books::CommentsController`,`Reports::CommentsController`を使うように修正。
