+++ app/controllers/reports/comments_controller.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+class Reports::CommentsController < CommentsController
+  before_action :set_commentable
+
+  private
+
+  def set_commentable
+    @commentable = Report.find(params[:report_id])
+  end
+
+  def render_commentable_show
+    @report = @commentable
+    render 'reports/show', status: :unprocessable_entity
+  end
+end
