+++ app/helpers/users_helper.rb
@@ -4,12 +4,4 @@ module UsersHelper
   def current_user_name(user)
     [user.name, user.email].find(&:present?)
   end
-
-  def name_or_email(resouce)
-    if resouce.user.name.strip.empty?
-      resouce.user.email
-    else
-      resouce.user.name
-    end
-  end
 end
