+++ app/views/books/_form.html.erb
@@ -1,15 +1,5 @@
 <%= form_with(model: book) do |form| %>
-  <% if book.errors.any? %>
-    <div style="color: red">
-      <h2><%= t('views.common.validation_error', errors: i18n_error_count(book.errors.count), name: Book.model_name.human.downcase) %>:</h2>
-
-      <ul>
-        <% book.errors.each do |error| %>
-          <li><%= error.full_message %></li>
-        <% end %>
-      </ul>
-    </div>
-  <% end %>
+  <%= render 'shared/form_error_messages', model: book %>
 
   <div>
     <%= form.label :title, style: "display: block" %>
