+++ app/views/reports/_form.html.erb
@@ -1,15 +1,5 @@
 <%= form_with(model: report) do |form| %>
-  <% if report.errors.any? %>
-    <div style="color: red">
-      <h2><%= t('views.common.validation_error', errors: i18n_error_count(report.errors.count), name: Report.model_name.human.downcase) %>:</h2>
-
-      <ul>
-        <% report.errors.each do |error| %>
-          <li><%= error.full_message %></li>
-        <% end %>
-      </ul>
-    </div>
-  <% end %>
+  <%= render 'shared/form_error_messages', model: report %>
 
   <div>
     <%= form.label :title, style: "display: block" %>
