+++ app/views/reports/show.html.erb
@@ -1,7 +1,7 @@
 <h1><%= t('views.common.title_show', name: Report.model_name.human) %></h1>
 
 <div class="show-item">
-  <%= render @report %>
+  <%= render 'reports/report', report: @report %>
 </div>
 
 <nav class="page-nav">
