+++ app/views/shared/_form_error_messages.html.erb
@@ -1,0 +1,11 @@
+<% if model.errors.any? %>
+  <div style="color: red">
+    <h2><%= t('views.common.validation_error', errors: i18n_error_count(model.errors.count), name: model.model_name.human.downcase) %></h2>
+
+    <ul>
+      <% model.errors.each do |error| %>
+        <li><%= error.full_message %></li>
+      <% end %>
+    </ul>
+  </div>
+<% end %>
