+++ config/routes.rb
@@ -2,8 +2,11 @@ Rails.application.routes.draw do
   mount LetterOpenerWeb::Engine, at: "/letter_opener" if Rails.env.development?
   devise_for :users
   root to: 'books#index'
-  resources :books, :reports do
-    resources :comments, only: %i(create destroy)
+  resources :books do
+    resources :comments, only: %i(create destroy), module: 'books'
+  end
+  resources :reports do
+    resources :comments, only: %i(create destroy), module: 'reports'
   end
   resources :users, only: %i(index show)
 end
